+++ src/node/tests/test_tests.py
@@ -52,17 +52,17 @@ class TestPatch(unittest.TestCase):
 
 class TestNodeTestCase(NodeTestCase):
 
-    def test_expect_error(self):
+    def test_expectError(self):
         def func_raises():
             raise Exception('Function raises')
-        err = self.expect_error(Exception, func_raises)
+        err = self.expectError(Exception, func_raises)
         self.assertEqual(str(err), 'Function raises')
 
         def func_passes():
             pass
         err = None
         try:
-            self.expect_error(Exception, func_passes)
+            self.expectError(Exception, func_passes)
         except Exception as e:
             err = e
         finally:
@@ -71,16 +71,16 @@ class TestNodeTestCase(NodeTestCase):
                 ' \'<function func_passes at '
             ) if IS_PY2 else (
                 'Expected \'<class \'Exception\'>\' when calling \'<function '
-                'TestNodeTestCase.test_expect_error.<locals>.func_passes at'
+                'TestNodeTestCase.test_expectError.<locals>.func_passes at'
             )
             self.assertTrue(str(err).startswith(expected))
 
-    def test_check_output(self):
+    def test_checkOutput(self):
         want = '...Hello...'
         got = 'Leading Hello Trailing'
-        self.check_output(want, got)
+        self.checkOutput(want, got)
         want = 'Hello'
-        err = self.expect_error(Failure, self.check_output, want, got)
+        err = self.expectError(Failure, self.checkOutput, want, got)
         self.assertEqual(str(err).split('\n'), [
             'Expected:',
             '    Hello',